deleted tactics, also known as attack patterns, from config collection
