+++ src/app/root-store/config/config.actions.ts
@@ -1,15 +1,10 @@
 import { Action } from '@ngrx/store';
 
-import { TacticChain } from '../../global/components/tactics-pane/tactics.model';
-import { Dictionary } from '../../models/json/dictionary';
-
 export const FETCH_CONFIG = '[Config] Fetch Config';
 export const ADD_CONFIG = '[Config] Add Config';
 export const UPDATE_CONFIG = '[Config] Update Config';
 export const DELETE_CONFIG = '[Config] Delete Config';
 export const CLEAR_CONFIG = '[Config] Clear Config';
-export const FETCH_TACTICS = '[Tactics] Fetch Tactics';
-export const LOAD_TACTICS = '[Tactics] Load Tactics';
 
 export class FetchConfig implements Action {
     public readonly type = FETCH_CONFIG;
@@ -37,21 +32,9 @@ export class ClearConfig implements Action {
     public readonly type = CLEAR_CONFIG;
 }
 
-export class FetchTactics implements Action {
-    public readonly type = FETCH_TACTICS;
-    constructor() {}
-}
-
-export class LoadTactics implements Action {
-    public readonly type = LOAD_TACTICS;
-    constructor(public payload: Dictionary<TacticChain>) {}
-}
-
 export type ConfigActions = 
     FetchConfig |
     AddConfig |
     UpdateConfig |
     DeleteConfig |
-    ClearConfig |
-    FetchTactics |
-    LoadTactics;
+    ClearConfig;
